tetragon: extract base sensor logic into a base sensor
We can treat the base set of programs and maps as a sensor. This create
a ./pkg/sensors/base to collect all this logic and pull it out of the
monolith sensors.
