+++ pkg/sensors/base/base.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Tetragon
 
-package sensors
+package base
 
 import (
 	"github.com/cilium/tetragon/pkg/kernels"
@@ -63,4 +63,39 @@ func GetExecveMap() *program.Map {
 		return ExecveMapV53
 	}
 	return ExecveMap
+}
+
+func GetDefaultPrograms() []*program.Program {
+	progs := []*program.Program{
+		Exit,
+		Fork,
+	}
+	if kernels.EnableLargeProgs() {
+		progs = append(progs, ExecveV53)
+	} else {
+		progs = append(progs, Execve)
+	}
+	return progs
+}
+
+func GetDefaultMaps() []*program.Map {
+	maps := []*program.Map{}
+
+	if kernels.EnableLargeProgs() {
+		maps = append(maps,
+			ExecveMapV53,
+			ExecveStatsV53,
+			NamesMapV53,
+			TCPMonMapV53,
+		)
+	} else {
+		maps = append(maps,
+			ExecveMap,
+			ExecveStats,
+			NamesMap,
+			TCPMonMap,
+		)
+	}
+	return maps
+
 }
